Update charging_recommendation.py
the function is tested and yields same result as before, given same parameter values in the json and the previous variables. However, the unit test files is designed with the variable inputs, therefore I did not change it as I am not certain of how the unit test file was coded.